This is an attempt at improving the Sophia implementation in HdtGraph (and also the implementation of Hdt).

The general idea is to use Arc<str> instead of MownStr, because Arc<str> can be cloned without actually copying the underlying text, and do not require a persistent "owner" of the text.

The performance gain is not obvious, because what takes most time is `FourSectDict::id_to_string`, i.e. the reconstruction of the string from the compressed dictionary. But it still seems that by using Arc<str> we spend less allocating memory than when using MownStr.

This PR also contains a number of minor improvements in the use of the Sophia API.