@KonradHoeffner 2023-02-22
# Benchmarks
## before

```
S??/1.1 (vincent, ?, ?) triple IDs
                        time:   [1.4519 µs 1.4540 µs 1.4545 µs]
S??/1.2 (vincent, ?, ?) str triples
                        time:   [3.2166 µs 3.2189 µs 3.2279 µs]
S??/1.3 (vincent, ?, ?) Sophia triples
                        time:   [4.4290 µs 4.4295 µs 4.4318 µs]

?P? 1429618 triples/2.1 (?, type, ?) triple IDs
                        time:   [342.98 ms 343.09 ms 343.53 ms]
?P? 1429618 triples/2.2 (?, type, ?) str triples
                        time:   [615.83 ms 617.06 ms 621.95 ms]
?P? 1429618 triples/2.3 (?, type, ?) Sophia triples
                        time:   [640.64 ms 640.78 ms 641.37 ms]

??O 1414214 triples/3.1 (?, ?, person) triple IDs
                        time:   [137.56 ms 137.71 ms 138.31 ms]
??O 1414214 triples/3.2 (?, ?, person) str triples
                        time:   [460.17 ms 460.19 ms 460.27 ms]
??O 1414214 triples/3.3 (?, ?, person) Sophia triples
                        time:   [496.97 ms 501.27 ms 502.35 ms]

?PO 1414214 triples/4.1 (?, type, person) triple IDs
                        time:   [117.41 ms 117.91 ms 118.04 ms]
?PO 1414214 triples/4.2 (?, type, person) str subjects
                        time:   [309.76 ms 310.24 ms 312.17 ms]
?PO 1414214 triples/4.3 (?, type, person) str triples
                        time:   [323.12 ms 323.46 ms 323.54 ms]
?PO 1414214 triples/4.4 (?, type, person) Sophia triples
                        time:   [357.56 ms 362.37 ms 363.57 ms]
```

## after

```
S??/1.1 (vincent, ?, ?) triple IDs
                        time:   [1.5132 µs 1.5139 µs 1.5168 µs]
                        change: [+4.0380% +4.2510% +4.4643%] (p = 0.10 > 0.05)
                        No change in performance detected.
S??/1.2 (vincent, ?, ?) str triples
                        time:   [6.2276 µs 6.2297 µs 6.2381 µs]
                        change: [+92.933% +93.432% +93.933%] (p = 0.10 > 0.05)
                        No change in performance detected.
S??/1.3 (vincent, ?, ?) Sophia triples
                        time:   [4.2669 µs 4.2670 µs 4.2670 µs]
                        change: [-3.7215% -3.6888% -3.6561%] (p = 0.07 > 0.05)
                        No change in performance detected.

?P? 1429618 triples/2.1 (?, type, ?) triple IDs
                        time:   [353.45 ms 353.56 ms 354.02 ms]
                        change: [+2.8878% +3.0539% +3.2204%] (p = 0.10 > 0.05)
                        No change in performance detected.
?P? 1429618 triples/2.2 (?, type, ?) str triples
                        time:   [654.45 ms 655.41 ms 659.26 ms]
                        change: [+5.2251% +6.1342% +7.0522%] (p = 0.10 > 0.05)
                        No change in performance detected.
?P? 1429618 triples/2.3 (?, type, ?) Sophia triples
                        time:   [715.80 ms 722.22 ms 723.83 ms]
                        change: [+11.603% +12.294% +12.986%] (p = 0.09 > 0.05)
                        No change in performance detected.

??O 1414214 triples/3.1 (?, ?, person) triple IDs
                        time:   [140.20 ms 140.24 ms 140.40 ms]
                        change: [+1.3715% +1.7188% +2.0680%] (p = 0.09 > 0.05)
                        No change in performance detected.
??O 1414214 triples/3.2 (?, ?, person) str triples
                        time:   [498.61 ms 498.84 ms 499.76 ms]
                        change: [+8.3285% +8.4660% +8.6034%] (p = 0.10 > 0.05)
                        No change in performance detected.
??O 1414214 triples/3.3 (?, ?, person) Sophia triples
                        time:   [552.20 ms 552.79 ms 552.93 ms]
                        change: [+9.9235% +10.589% +11.261%] (p = 0.10 > 0.05)
                        No change in performance detected.

?PO 1414214 triples/4.1 (?, type, person) triple IDs
                        time:   [118.90 ms 118.90 ms 118.91 ms]
                        change: [+0.7300% +1.0019% +1.2754%] (p = 0.08 > 0.05)
                        No change in performance detected.
?PO 1414214 triples/4.2 (?, type, person) str subjects
                        time:   [313.29 ms 313.74 ms 313.86 ms]
                        change: [+0.3570% +0.8375% +1.3217%] (p = 0.12 > 0.05)
                        No change in performance detected.
?PO 1414214 triples/4.3 (?, type, person) str triples
                        time:   [338.70 ms 338.86 ms 339.47 ms]
                        change: [+4.6851% +4.8734% +5.0619%] (p = 0.10 > 0.05)
                        No change in performance detected.
?PO 1414214 triples/4.4 (?, type, person) Sophia triples
                        time:   [360.62 ms 360.67 ms 360.68 ms]
                        change: [-0.8126% +0.0232% +0.8730%] (p = 0.82 > 0.05)
                        No change in performance detected.
```