@KonradHoeffner 2023-02-22
# Benchmarking ??? pattern (all triples) on i9-10900k

## Before

```
Benchmarking ??? (all)/0.1 all triple IDs: Warming up for 3.0000 s
??? (all)/0.1 all triple IDs
                        time:   [1.6171 s 1.6180 s 1.6189 s]
??? (all)/0.2 all str triples
                        time:   [6.3724 s 6.3750 s 6.3785 s]
??? (all)/0.3 all Sophia triples
                        time:   [7.8375 s 7.8492 s 7.8624 s]
```

### After

```
??? (all)/0.1 all triple IDs
                        time:   [1.6182 s 1.6336 s 1.6480 s]
                        change: [+0.1297% +0.9638% +1.8168%] (p = 0.07 > 0.05)
                        No change in performance detected.
??? (all)/0.2 all str triples
                        time:   [7.1611 s 7.2483 s 7.3353 s]
                        change: [+12.424% +13.699% +14.932%] (p = 0.00 < 0.05)
                        Performance has regressed.
??? (all)/0.3 all Sophia triples
                        time:   [8.5335 s 8.5736 s 8.6102 s]
                        change: [+8.7127% +9.2289% +9.6978%] (p = 0.00 < 0.05)
                        Performance has regressed.
```