@KonradHoeffner 2023-02-22
Maybe the idx variable needs to be updated in the else part of the get_x_string function?