@KonradHoeffner 2023-02-22
```
    fn get_x_string(&mut self, i: usize, pos: usize, kind: &'static IdKind) -> Result<Arc<str>, DictErr> {
        debug_assert!(i != 0);
        if self.idx[pos] == i {
            Ok(self.arc[pos].as_ref().unwrap().clone())
        } else {
            let ret: Arc<str> = self.hdt.dict.id_to_string(i, kind)?.into();
            self.arc[pos] = Some(ret.clone());
+          self.idx[pos] = i;
            Ok(ret)
        }
```

```
??? (all)/0.1 all triple IDs
                        time:   [1.6308 s 1.6329 s 1.6366 s]
                        change: [+0.7702% +0.9257% +1.1700%] (p = 0.00 < 0.05)
                        Change within noise threshold.
??? (all)/0.2 all str triples
                        time:   [5.8885 s 5.8956 s 5.9026 s]
                        change: [-7.6339% -7.5201% -7.3964%] (p = 0.00 < 0.05)
                        Performance has improved.
??? (all)/0.3 all Sophia triples
                        time:   [6.9292 s 6.9454 s 6.9653 s]
                        change: [-11.755% -11.514% -11.252%] (p = 0.00 < 0.05)
                        Performance has improved.
```