@pchampin 2023-02-24
> Maybe the idx variable needs to be updated in the else part of the get_x_string function?

Duh! So silly of me... :facepalm: Thanks for fixing this.